Implements Node/Assignment(.cache).read/list
Adds the idea of a cached read/list, which is synchronous.
This is potentially a problematic way to implement it, but for the time
being it saves some effort until it can be addressed properly with a
solution that has a single API and consults the cache intelligently,
falling back to the server where required.