Use event buffer for accounting for node resuming
This 'fixes' (for lack of a better word) some issues around race
conditions when attempting to account for the resuming of a node outside
of the buffered event stream. By inserting a new event into the buffered
stream, it effects the change in the tabIdMap outside the execution of
any other event handler.