Implements TAB_UPDATED
The implementation is large and unwieldy at this point, but will suffice
for now.
Ideally we could use Treo and promise chaining, however there remains
the possibility that data could change between read and update. Not sure
whether this is a potential issue or not at this point, so I'm erring on
the side of caution and using the lower level IDB Transaction API, so no
chaining/filtering.